Commit by Valeria, changed Person and created Test with inner Class Question
