+++ src/main/java/core/entity/Person.java
@@ -26,32 +26,16 @@ public class Person {
         return id;
     }
 
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
     public String getFirstName() {
         return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
     }
 
     public String getLastName() {
         return lastName;
     }
 
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
     public String geteMail() {
         return eMail;
-    }
-
-    public void seteMail(String eMail) {
-        this.eMail = eMail;
     }
 
     public String getCourseName() {
